+++ tests/fixtures.py
@@ -1,0 +1,8 @@
+import pytest
+
+@pytest.fixture
+def paragraphs_type_exclude():
+    return [
+        "microsoft word",
+        "uoa",
+    ]